returnallcategories

+++ Assignment4/OrderDetails.cs
@@ -6,8 +6,8 @@ namespace Assignment4
         public Order Order { get; set; }
         public int ProductId { get; set; }
         public Product Product { get; set; }
-        public double UnitPrice { get; set; }
-        public double Quantity { get; set; }
-        public double Discount { get; set; }
+        public int UnitPrice { get; set; }
+        public int Quantity { get; set; }
+        public int Discount { get; set; }
     }
 }
